requirements.txt
python-dotenv
pyodbc
faker
